requirements.txt
discord-py-slash-command
python-dotenv
discord-py-interactions
async_timeout